feat: Boiler-plate code for cookie prompt
Added the initial code needed for displaying the cookie prompt to users. The package react-cookie is
installed in this commit and used in the code, where the cookie is read to make a decision whether
the prompt should be displayed or not. The prompt is a simple div with an experimental message
currently.