Added sass rewiring
In this commit, a rewiring package was added to the project. This allows us to define custom configurations for webpack without ejecting the project. We used a css module rewiring package, which allows us to include sass files inside the react components, as well as using css modules within the react components