Added the use of CSS modules
In this commit, we introduce the use of CSS modules. CSS modules are a way to modularize CSS classes, which prevents scope creep to a large extent. The current components were refactored to take into use the CSS modules, as such some of the UI elements might not be optimized.